refactor: move highlight words module
